Modified TestRunHook validation scenario to include user-provided Before/After TestRun hooks. AfterTestRun hooks are properly handled; Before are not.
Refactored the CucumberMessagePublisher logic in how it handles hookBindingStarted and Finished execution events to clean up the code and move envelope creation responsibility to the hook tracker. This should make it easier to further refactor should we choose to change how we emit messages for the Before/After Feature hooks which don't map cleanly to the Cucumber Messages spec.