Refactor unit test generation for clarity and structure per review comments.
This commit introduces a new class `UnitTestFeatureGenerationResult` to encapsulate the results of generating unit test fixtures, replacing out parameters with a more structured return type.
The naming of variables that contain Messages has been adjusted to adopt a convention of avoiding the term 'ndjson', and the `FeatureNdjsonMessages` property has been renamed to `FeatureMessages` for consistency.

Output from the Generation Task has been modified such that the feature messages and the resource name we'd like them to be embedded at are now Metadata items on the returned ITaskItem.