Embed all Feature Level Messages into the Assembly at compile time (#791)
This PR changes the way that Feature-level Messages (aka, Source, GherkinDocument, and Pickles) are handled at code generation and runtime.

These messages are now generated at code-generation time and serialized to a file which is then embedded into the compiled assembly.
At test run-time, these are read from the assembly and published.

This change allows us to support test runs of very large features (which previously caused a compiler error).
It also has the benefit of improving overall performance when Formatters are enabled.

To support this change, changes were made to the data structure (FeatureLevelCucumberMessages) which provides to the runtime the Feature-level Messages.
Changes were also made to the Reqnroll.Generator and Reqnroll.Tools.MsBuild.Generation code to convey the generated messages from the code-generator to the .targets file.
---------

Co-authored-by: Gáspár Nagy <gaspar.nagy@gmail.com>