Changed the default BuildDataData record so that the ProductName property is empty, rather than "UNKNOWN". (#800)
* Changed the default BuildDataData record so that the ProductName property is empty, rather than "UNKNOWN".
This causes the Build Info banner in the HTML report to be suppressed (rather than display UNKNOWN for every sub-property).

* BuildData BuildServer Product should be null when no Build Server can be determined

This change will force the HTML reports to suppress the Build Info banner at the top of HTML reports when the Build Server cannot be determined (rather than displaying UNKNOWN for all segments of the banner).