RNR#883 AfterScenario hooks are not executed for Ignored, Pending, or Inconclusive statuses. (#900)
* RNR#883 AfterScenario hooks are not executed for Ignored, Pending, or Inconclusive statuses.
This commit adds a test to the Hook Error Handling Spec test.
This commit assumes that current behavior is correct, but can easily be changed if it is agreed that the AfterScenario hooks should fire under those statuses.
This gives us a statement of intention and a platform for proving correct behavior of any changes we might make to fire the AfterScenario events under those circumstances.

* Revised logic to accomodate dynamically skipped scenarios.
Revised the Hooks feature in the Specs to match the desired outcome (passing).

* Update CHANGELOG.md

* add unit test

---------

Co-authored-by: Gáspár Nagy <gaspar.nagy@gmail.com>