Added support for variable substitution within a configured Formatter's outputFilePath string (#930)
* Added support for variable substitution within a configured Formatter's outputFilePath string. Supported substitutions include {timestamp} and {environmentVariable}.

* Template based substitution logic moved to a new service: VariableSubstitutionService.
Added feature to substitute in 'buildNumber', 'branch', 'revision', and 'tag' as known substitution template items. Data for these pulled from the BuildMetadataProvider.
For now, this is invoked from the FormattersConfigurationProvider, a dependency already of the FileWrittingFormatterBase class. (We can later make the VariableSubstitutionService a direct dependency of the FileWritingFormatterBase)
Modified Html and Messages formatters to eliminate the impact to their constructors.
Updated tests affected.
Moved unit tests to dedicated test class (VariableSubstitutionServiceTests).

* Remove unnecessary usings.

* Review comment cleanups.

* Draft of documentation for Variable Substitution